move insights behind the feature flag
