+++ packages/compass-collection/src/components/collection/collection.tsx
@@ -3,6 +3,7 @@ import { createLoggerAndTelemetry } from '@mongodb-js/compass-logging';
 import type { Document } from 'mongodb';
 import React, { useCallback, useEffect, useMemo } from 'react';
 import { TabNavBar, css } from '@mongodb-js/compass-components';
+import type { Signal } from '@mongodb-js/compass-components';
 import { usePreference } from 'compass-preferences-model';
 import CollectionHeader from '../collection-header';
 import { getCollectionStatsInitialState } from '../../modules/stats';
@@ -32,6 +33,52 @@ const collectionContainerStyles = css({
 const collectionModalContainerStyles = css({
   zIndex: 100,
 });
+
+const getInsightsForPipeline = (pipeline: Document[], isAtlas: boolean) => {
+  const insights: Record<string, Signal> = {};
+  for (const stage of pipeline) {
+    if ('$lookup' in stage) {
+      const signal = {
+        id: 'lookup-usage-in-view',
+        title: 'This view uses $lookup',
+        description:
+          '$lookup operations can be resource intensive because they perform operations on two collections instead of one. Consider embedding documents or arrays to increase read performance.',
+        learnMoreLink:
+          'https://www.mongodb.com/docs/atlas/schema-suggestions/reduce-lookup-operations/#std-label-anti-pattern-denormalization',
+      };
+      insights[signal.id] = signal;
+    }
+
+    if ('$match' in stage) {
+      const stringifiedStageValue = JSON.stringify(stage);
+      if (
+        stringifiedStageValue.includes('$regex') ||
+        stringifiedStageValue.includes('$text')
+      ) {
+        const signal = isAtlas
+          ? {
+              id: 'atlas-text-regex-usage-in-view',
+              title: 'This view uses an inefficient text search operator',
+              description:
+                "In many cases, Atlas Search is MongoDB's most efficient full text search option. Convert your view’s query to $search for a wider range of functionality.",
+              learnMoreLink:
+                'https://www.mongodb.com/docs/atlas/atlas-search/best-practices/',
+            }
+          : {
+              id: 'non-atlas-text-regex-usage-in-view',
+              title: 'This view uses an inefficient text search operator',
+              description:
+                "In many cases, Atlas Search is MongoDB's most efficient full text search option. Connect with Atlas to explore the power of Atlas Search.",
+              learnMoreLink:
+                'https://www.mongodb.com/docs/atlas/atlas-search/best-practices/',
+            };
+
+        insights[signal.id] = signal;
+      }
+    }
+  }
+  return Object.values(insights);
+};
 
 type CollectionProps = {
   darkMode?: boolean;
@@ -140,6 +187,14 @@ const Collection: React.FunctionComponent<CollectionProps> = ({
     };
   }, [localAppRegistry, onSubTabClicked, tabs]);
 
+  const showInsights = usePreference('showInsights', React);
+
+  const insights = useMemo(() => {
+    return showInsights && pipeline?.length
+      ? getInsightsForPipeline(pipeline, isAtlas)
+      : [];
+  }, [pipeline, isAtlas, showInsights]);
+
   return (
     <div className={collectionStyles} data-testid="collection">
       <div className={collectionContainerStyles}>
@@ -157,7 +212,7 @@ const Collection: React.FunctionComponent<CollectionProps> = ({
           pipeline={pipeline}
           sourceName={sourceName}
           stats={stats[namespace] ?? getCollectionStatsInitialState()}
-          isAtlas={isAtlas}
+          insights={insights}
         />
         <TabNavBar
           data-testid="collection-tabs"
